Minor fixes in check/install hooks commands

+++ src/Command/InstallCommand.php
@@ -95,6 +95,8 @@ class InstallCommand extends Command
             }
         }
 
-        rmdir($directory);
+        if (is_dir($directory)) {
+            rmdir($directory);
+        }
     }
 }
